pyproject.toml
[project]
name = "api-nginx-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.233.0",
    "boto3>=1.42.17",
    "constructs>=10.4.4",
    "fastapi>=0.127.0",
    "gunicorn>=23.0.0",
    "mlflow>=3.8.1",
    "uvicorn>=0.40.0",
]
